# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Planificacion Horaria",
        "description": "Gestion de horarios de clases",
        "contact": {
            "email": "contacto@tudominio.com"
        },
        "license": {
            "name": "Licencia",
            "url": "https://www.terciarioUrquiza.com.ar/planificacionHoraria"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/materias": {
            "get": {
                "tags": [
                    "Materias"
                ],
                "summary": "Obtener todas las materias",
                "operationId": "918cb2607842cfe87ff9f6796c4cc540",
                "responses": {
                    "200": {
                        "description": "Obtener todas las materias"
                    },
                    "500": {
                        "description": "Error al obtener las materias"
                    }
                }
            }
        },
        "/api/materias/{id}": {
            "get": {
                "tags": [
                    "Materias"
                ],
                "summary": "Obtener una materia por id",
                "operationId": "808b887ba424ec0048dbd9a11aa22ffe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la materia",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Obtener una materia por id"
                    },
                    "404": {
                        "description": "No existe la materia"
                    },
                    "500": {
                        "description": "Error al obtener la materia"
                    }
                }
            }
        },
        "/api/materias/nombre/{nombre}": {
            "get": {
                "tags": [
                    "Materias"
                ],
                "summary": "Obtener una materia por nombre",
                "operationId": "77797f3b259f8f181c971a1a38a5b35b",
                "parameters": [
                    {
                        "name": "nombre",
                        "in": "path",
                        "description": "Nombre de la materia",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "int64"
                        },
                        "example": "Programacion"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Obtener una materia por nombre",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/MateriaValidate"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No existe la materia"
                    },
                    "500": {
                        "description": "Error al obtener la materia"
                    }
                }
            }
        },
        "/api/materias/guardar": {
            "post": {
                "tags": [
                    "Materias"
                ],
                "summary": "Guardar una materia",
                "operationId": "03f231d6d3580d529579faf20ce0207e",
                "requestBody": {
                    "description": "Datos de la materia",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MateriaData"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Guardar una materia"
                    },
                    "400": {
                        "description": "Ya existe una materia con ese nombre"
                    },
                    "500": {
                        "description": "Error al guardar la materia"
                    }
                }
            }
        },
        "/api/materias/actualizar/{id}": {
            "put": {
                "tags": [
                    "Materias"
                ],
                "summary": "Actualizar una materia",
                "operationId": "0682bdb84e7fe75b40f6ba6d74fccb24",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la materia",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "description": "Datos de la materia",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MateriaData"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Actualizar una materia"
                    },
                    "404": {
                        "description": "No existe la materia"
                    },
                    "500": {
                        "description": "Error al actualizar la materia"
                    }
                }
            }
        },
        "/api/materias/eliminar/{id}": {
            "delete": {
                "tags": [
                    "Materias"
                ],
                "summary": "Eliminar una materia",
                "operationId": "0a98ca976cdd4d9082492a0b78c3f495",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la materia",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Eliminar una materia"
                    },
                    "404": {
                        "description": "No existe la materia"
                    },
                    "500": {
                        "description": "Error al eliminar la materia"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "MateriaData": {
                "title": "MateriaData",
                "description": "Esquema del objeto MateriaData",
                "properties": {
                    "nombre": {
                        "description": "Nombre de la materia",
                        "type": "string",
                        "example": "Matematicas"
                    }
                },
                "type": "object"
            },
...
            "MateriaValidate": {
                "title": "MateriaValidate",
                "description": "Esquema del objeto MateriaValidate",
                "properties": {
                    "nombre": {
                        "description": "Nombre de la materia",
                        "type": "string"
                    },
                    "fecha_creacion": {
                        "description": "Fecha de creacion de la materia",
                        "type": "string",
                        "format": "date-time"
                    },
                    "fecha_modificacion": {
                        "description": "Fecha de modificacion de la materia",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    }
}